order_service/order_service/requirements.txt
celery==5.4.0
Django==4.2.16
django-celery-results==2.5.1
django-fsm==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2
gunicorn==23.0.0
pika
